GameController 클래스 추가
전반적인 키보드 입력을 관리하는GameController를 추가하여 GameMgr이 상속받아 키보드로 사각형을 이동하게 만들었다.

+++ SimpleGame/SimpleGame/stdafx.h
@@ -21,5 +21,5 @@ constexpr float meter(float cMeter = 1)
 	return cMeter * 100;
 }
 
-constexpr float wndSizeX{ meter(5) };
-constexpr float wndSizeY{ meter(5) };
+constexpr int wndSizeX{ static_cast<int>(meter(5)) };
+constexpr int wndSizeY{ static_cast<int>(meter(5)) };
